Revert "App backscaling on demand (#458)"
This reverts commit e5e73fd6c585ee18adde3f3b3b1dcc6e79af35e9.
