+++ config/crds/migration_v1alpha1_migmigration.yaml
@@ -49,8 +49,6 @@ spec:
               items:
                 type: string
               type: array
-            itenerary:
-              type: string
             observedDigest:
               type: string
             phase:
