+++ pkg/apis/migration/v1alpha1/labels.go
@@ -1,10 +1,9 @@
 package v1alpha1
 
 import (
-	"strings"
-
 	migref "github.com/konveyor/mig-controller/pkg/reference"
 	"k8s.io/apimachinery/pkg/types"
+	"strings"
 )
 
 // Labels
