+++ pkg/apis/migration/v1alpha1/model.go
@@ -2,12 +2,11 @@ package v1alpha1
 
 import (
 	"context"
-	"strconv"
-
 	kapi "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/types"
 	k8sclient "sigs.k8s.io/controller-runtime/pkg/client"
+	"strconv"
 )
 
 //
