+++ pkg/compat/client.go
@@ -21,7 +21,6 @@ import (
 // A smart client.
 // Provides seamless API version compatibility.
 type Client struct {
-	*rest.Config
 	k8sclient.Client
 	dapi.DiscoveryInterface
 	// major k8s version.
@@ -58,7 +57,6 @@ func NewClient(restCfg *rest.Config) (k8sclient.Client, error) {
 		return nil, err
 	}
 	nClient := &Client{
-		Config:             restCfg,
 		Client:             rClient,
 		DiscoveryInterface: dClient,
 		Major:              major,
