+++ pkg/controller/migmigration/migmigration_controller_test.go
@@ -17,11 +17,10 @@ limitations under the License.
 package migmigration
 
 import (
+	"github.com/onsi/gomega"
 	"reflect"
 	"testing"
 	"time"
-
-	"github.com/onsi/gomega"
 
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -87,23 +86,26 @@ func TestReconcile(t *testing.T) {
 // but for not they are expected to be the identical.
 func Test_Itineraries(t *testing.T) {
 	g := gomega.NewGomegaWithT(t)
-	stage := StageItinerary
-	common := Itinerary{}
+	stage := StageItinerary[4 : len(StageItinerary)-1]
 
-	for i, step := range StageItinerary.Steps {
-		found := false
-		for _, finalStep := range FinalItinerary.Steps[i:] {
-			if step.phase == finalStep.phase {
-				common.Steps = append(common.Steps, step)
-				found = true
-				break
-			}
-		}
-		if !found {
-			t.Errorf("'%s' is not contained within the final itinerary", step.phase)
+	begin := 0
+	for i, step := range FinalItinerary {
+		if step.phase == stage[0].phase {
+			begin = i
+			break
 		}
 	}
+	end := 0
+	last := stage[len(stage)-1]
+	for i, step := range FinalItinerary {
+		if step.phase == last.phase {
+			end = i + 1
+			break
+		}
+	}
+	final := FinalItinerary[begin:end]
 
-	g.Expect(reflect.DeepEqual(stage.Steps, common.Steps)).To(gomega.BeTrue())
-
+	g.Expect(begin == 0).To(gomega.BeFalse())
+	g.Expect(end < len(FinalItinerary)).To(gomega.BeTrue())
+	g.Expect(reflect.DeepEqual(stage, final)).To(gomega.BeTrue())
 }
