+++ pkg/controller/migmigration/migrate.go
@@ -83,7 +83,6 @@ func (r *ReconcileMigMigration) migrate(migration *migapi.MigMigration) (time.Du
 
 	// Result
 	migration.Status.Phase = task.Phase
-	migration.Status.Itenerary = task.Itinerary.Name
 
 	// Completed
 	if task.Phase == Completed {
