+++ pkg/controller/migmigration/quiesce.go
@@ -2,8 +2,6 @@ package migmigration
 
 import (
 	"context"
-	"encoding/json"
-	"strconv"
 
 	ocappsv1 "github.com/openshift/api/apps/v1"
 	appsv1 "k8s.io/api/apps/v1"
@@ -21,83 +19,37 @@ func (t *Task) quiesceApplications() error {
 		log.Trace(err)
 		return err
 	}
-	err = t.quiesceCronJobs(client)
+	err = t.suspendCronJobs(client)
 	if err != nil {
 		log.Trace(err)
 		return err
 	}
-	err = t.quiesceDeploymentConfigs(client)
+	err = t.scaleDownDeploymentConfigs(client)
 	if err != nil {
 		log.Trace(err)
 		return err
 	}
-	err = t.quiesceDeployments(client)
+	err = t.scaleDownDeployments(client)
 	if err != nil {
 		log.Trace(err)
 		return err
 	}
-	err = t.quiesceStatefulSets(client)
+	err = t.scaleDownStatefulSets(client)
 	if err != nil {
 		log.Trace(err)
 		return err
 	}
-	err = t.quiesceReplicaSets(client)
+	err = t.scaleDownReplicaSets(client)
 	if err != nil {
 		log.Trace(err)
 		return err
 	}
-	err = t.quiesceDaemonSets(client)
+	err = t.scaleDownDaemonSets(client)
 	if err != nil {
 		log.Trace(err)
 		return err
 	}
-	err = t.quiesceJobs(client)
-	if err != nil {
-		log.Trace(err)
-		return err
-	}
-
-	return nil
-}
-
-// All quiesce undo functionality should be put here
-func (t *Task) unQuiesceApplications() error {
-	client, err := t.getSourceClient()
-	if err != nil {
-		log.Trace(err)
-		return err
-	}
-	err = t.unQuiesceCronJobs(client)
-	if err != nil {
-		log.Trace(err)
-		return err
-	}
-	err = t.unQuiesceDeploymentConfigs(client)
-	if err != nil {
-		log.Trace(err)
-		return err
-	}
-	err = t.unQuiesceDeployments(client)
-	if err != nil {
-		log.Trace(err)
-		return err
-	}
-	err = t.unQuiesceStatefulSets(client)
-	if err != nil {
-		log.Trace(err)
-		return err
-	}
-	err = t.unQuiesceReplicaSets(client)
-	if err != nil {
-		log.Trace(err)
-		return err
-	}
-	err = t.unQuiesceDaemonSets(client)
-	if err != nil {
-		log.Trace(err)
-		return err
-	}
-	err = t.unQuiesceJobs(client)
+	err = t.scaleDownJobs(client)
 	if err != nil {
 		log.Trace(err)
 		return err
@@ -107,7 +59,7 @@ func (t *Task) unQuiesceApplications() error {
 }
 
 // Scales down DeploymentConfig on source cluster
-func (t *Task) quiesceDeploymentConfigs(client k8sclient.Client) error {
+func (t *Task) scaleDownDeploymentConfigs(client k8sclient.Client) error {
 	for _, ns := range t.sourceNamespaces() {
 		list := ocappsv1.DeploymentConfigList{}
 		options := k8sclient.InNamespace(ns)
@@ -120,13 +72,9 @@ func (t *Task) quiesceDeploymentConfigs(client k8sclient.Client) error {
 			return err
 		}
 		for _, dc := range list.Items {
-			if dc.Annotations == nil {
-				dc.Annotations = make(map[string]string)
-			}
 			if dc.Spec.Replicas == 0 {
 				continue
 			}
-			dc.Annotations[ReplicasAnnotation] = strconv.FormatInt(int64(dc.Spec.Replicas), 10)
 			dc.Spec.Replicas = 0
 			err = client.Update(context.TODO(), &dc)
 			if err != nil {
@@ -139,47 +87,8 @@ func (t *Task) quiesceDeploymentConfigs(client k8sclient.Client) error {
 	return nil
 }
 
-// Scales DeploymentConfig back up on source cluster
-func (t *Task) unQuiesceDeploymentConfigs(client k8sclient.Client) error {
-	for _, ns := range t.sourceNamespaces() {
-		list := ocappsv1.DeploymentConfigList{}
-		options := k8sclient.InNamespace(ns)
-		err := client.List(
-			context.TODO(),
-			options,
-			&list)
-		if err != nil {
-			log.Trace(err)
-			return err
-		}
-		for _, dc := range list.Items {
-			if dc.Annotations == nil {
-				continue
-			}
-			replicas, exist := dc.Annotations[ReplicasAnnotation]
-			if !exist {
-				continue
-			}
-			number, err := strconv.Atoi(replicas)
-			if err != nil {
-				log.Trace(err)
-				return err
-			}
-			delete(dc.Annotations, ReplicasAnnotation)
-			dc.Spec.Replicas = int32(number)
-			err = client.Update(context.TODO(), &dc)
-			if err != nil {
-				log.Trace(err)
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
 // Scales down all Deployments
-func (t *Task) quiesceDeployments(client k8sclient.Client) error {
+func (t *Task) scaleDownDeployments(client k8sclient.Client) error {
 	zero := int32(0)
 	for _, ns := range t.sourceNamespaces() {
 		list := appsv1.DeploymentList{}
@@ -193,13 +102,9 @@ func (t *Task) quiesceDeployments(client k8sclient.Client) error {
 			return err
 		}
 		for _, deployment := range list.Items {
-			if deployment.Annotations == nil {
-				deployment.Annotations = make(map[string]string)
-			}
-			if *deployment.Spec.Replicas == zero {
+			if deployment.Spec.Replicas == &zero {
 				continue
 			}
-			deployment.Annotations[ReplicasAnnotation] = strconv.FormatInt(int64(*deployment.Spec.Replicas), 10)
 			deployment.Spec.Replicas = &zero
 			err = client.Update(context.TODO(), &deployment)
 			if err != nil {
@@ -212,48 +117,8 @@ func (t *Task) quiesceDeployments(client k8sclient.Client) error {
 	return nil
 }
 
-// Scales all Deployments back up
-func (t *Task) unQuiesceDeployments(client k8sclient.Client) error {
-	for _, ns := range t.sourceNamespaces() {
-		list := appsv1.DeploymentList{}
-		options := k8sclient.InNamespace(ns)
-		err := client.List(
-			context.TODO(),
-			options,
-			&list)
-		if err != nil {
-			log.Trace(err)
-			return err
-		}
-		for _, deployment := range list.Items {
-			if deployment.Annotations == nil {
-				deployment.Annotations = make(map[string]string)
-			}
-			replicas, exist := deployment.Annotations[ReplicasAnnotation]
-			if !exist {
-				continue
-			}
-			number, err := strconv.Atoi(replicas)
-			if err != nil {
-				log.Trace(err)
-				return err
-			}
-			delete(deployment.Annotations, ReplicasAnnotation)
-			restoredReplicas := int32(number)
-			deployment.Spec.Replicas = &restoredReplicas
-			err = client.Update(context.TODO(), &deployment)
-			if err != nil {
-				log.Trace(err)
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
 // Scales down all StatefulSets.
-func (t *Task) quiesceStatefulSets(client k8sclient.Client) error {
+func (t *Task) scaleDownStatefulSets(client k8sclient.Client) error {
 	zero := int32(0)
 	for _, ns := range t.sourceNamespaces() {
 		list := appsv1.StatefulSetList{}
@@ -267,53 +132,10 @@ func (t *Task) quiesceStatefulSets(client k8sclient.Client) error {
 			return err
 		}
 		for _, set := range list.Items {
-			if set.Annotations == nil {
-				set.Annotations = make(map[string]string)
-			}
-			if *set.Spec.Replicas == zero {
+			if set.Spec.Replicas == &zero {
 				continue
 			}
-			set.Annotations[ReplicasAnnotation] = strconv.FormatInt(int64(*set.Spec.Replicas), 10)
 			set.Spec.Replicas = &zero
-			err = client.Update(context.TODO(), &set)
-			if err != nil {
-				log.Trace(err)
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// Scales all StatefulSets back up
-func (t *Task) unQuiesceStatefulSets(client k8sclient.Client) error {
-	for _, ns := range t.sourceNamespaces() {
-		list := appsv1.StatefulSetList{}
-		options := k8sclient.InNamespace(ns)
-		err := client.List(
-			context.TODO(),
-			options,
-			&list)
-		if err != nil {
-			log.Trace(err)
-			return err
-		}
-		for _, set := range list.Items {
-			if set.Annotations == nil {
-				continue
-			}
-			replicas, exist := set.Annotations[ReplicasAnnotation]
-			if !exist {
-				continue
-			}
-			number, err := strconv.Atoi(replicas)
-			if err != nil {
-				log.Trace(err)
-				return err
-			}
-			delete(set.Annotations, ReplicasAnnotation)
-			restoredReplicas := int32(number)
-			set.Spec.Replicas = &restoredReplicas
 			err = client.Update(context.TODO(), &set)
 			if err != nil {
 				log.Trace(err)
@@ -325,7 +147,7 @@ func (t *Task) unQuiesceStatefulSets(client k8sclient.Client) error {
 }
 
 // Scales down all ReplicaSets.
-func (t *Task) quiesceReplicaSets(client k8sclient.Client) error {
+func (t *Task) scaleDownReplicaSets(client k8sclient.Client) error {
 	zero := int32(0)
 	for _, ns := range t.sourceNamespaces() {
 		list := appsv1.ReplicaSetList{}
@@ -339,13 +161,9 @@ func (t *Task) quiesceReplicaSets(client k8sclient.Client) error {
 			return err
 		}
 		for _, set := range list.Items {
-			if set.Annotations == nil {
-				set.Annotations = make(map[string]string)
-			}
-			if *set.Spec.Replicas == zero {
+			if set.Spec.Replicas == &zero {
 				continue
 			}
-			set.Annotations[ReplicasAnnotation] = strconv.FormatInt(int64(*set.Spec.Replicas), 10)
 			set.Spec.Replicas = &zero
 			err = client.Update(context.TODO(), &set)
 			if err != nil {
@@ -357,47 +175,14 @@ func (t *Task) quiesceReplicaSets(client k8sclient.Client) error {
 	return nil
 }
 
-// Scales all ReplicaSets back up
-func (t *Task) unQuiesceReplicaSets(client k8sclient.Client) error {
-	for _, ns := range t.sourceNamespaces() {
-		list := appsv1.ReplicaSetList{}
-		options := k8sclient.InNamespace(ns)
-		err := client.List(
-			context.TODO(),
-			options,
-			&list)
-		if err != nil {
-			log.Trace(err)
-			return err
-		}
-		for _, set := range list.Items {
-			if set.Annotations == nil {
-				continue
-			}
-			replicas, exist := set.Annotations[ReplicasAnnotation]
-			if !exist {
-				continue
-			}
-			number, err := strconv.Atoi(replicas)
-			if err != nil {
-				log.Trace(err)
-				return err
-			}
-			delete(set.Annotations, ReplicasAnnotation)
-			restoredReplicas := int32(number)
-			set.Spec.Replicas = &restoredReplicas
-			err = client.Update(context.TODO(), &set)
-			if err != nil {
-				log.Trace(err)
-				return err
-			}
-		}
-	}
-	return nil
-}
+// values to support nodeSelector-based DaemonSet quiesce
+const (
+	quiesceNodeSelector    = "openshift.io/quiesceDaemonSet"
+	quiesceNodeSelectorVal = "quiesce"
+)
 
 // Scales down all DaemonSets.
-func (t *Task) quiesceDaemonSets(client k8sclient.Client) error {
+func (t *Task) scaleDownDaemonSets(client k8sclient.Client) error {
 	for _, ns := range t.sourceNamespaces() {
 		list := appsv1.DaemonSetList{}
 		options := k8sclient.InNamespace(ns)
@@ -410,60 +195,12 @@ func (t *Task) quiesceDaemonSets(client k8sclient.Client) error {
 			return err
 		}
 		for _, set := range list.Items {
-			if set.Annotations == nil {
-				set.Annotations = make(map[string]string)
-			}
 			if set.Spec.Template.Spec.NodeSelector == nil {
-				set.Spec.Template.Spec.NodeSelector = map[string]string{}
-			} else if _, exist := set.Spec.Template.Spec.NodeSelector[QuiesceNodeSelector]; exist {
+				set.Spec.Template.Spec.NodeSelector = make(map[string]string)
+			} else if set.Spec.Template.Spec.NodeSelector[quiesceNodeSelector] == quiesceNodeSelectorVal {
 				continue
 			}
-			selector, err := json.Marshal(set.Spec.Template.Spec.NodeSelector)
-			if err != nil {
-				log.Trace(err)
-				return err
-			}
-			set.Annotations[NodeSelectorAnnotation] = string(selector)
-			set.Spec.Template.Spec.NodeSelector[QuiesceNodeSelector] = "true"
-			err = client.Update(context.TODO(), &set)
-			if err != nil {
-				log.Trace(err)
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// Scales all DaemonSets back up
-func (t *Task) unQuiesceDaemonSets(client k8sclient.Client) error {
-	for _, ns := range t.sourceNamespaces() {
-		list := appsv1.DaemonSetList{}
-		options := k8sclient.InNamespace(ns)
-		err := client.List(
-			context.TODO(),
-			options,
-			&list)
-		if err != nil {
-			log.Trace(err)
-			return err
-		}
-		for _, set := range list.Items {
-			if set.Annotations == nil {
-				continue
-			}
-			selector, exist := set.Annotations[NodeSelectorAnnotation]
-			if !exist {
-				continue
-			}
-			nodeSelector := map[string]string{}
-			err := json.Unmarshal([]byte(selector), &nodeSelector)
-			if err != nil {
-				log.Trace(err)
-				return err
-			}
-			delete(set.Annotations, NodeSelectorAnnotation)
-			set.Spec.Template.Spec.NodeSelector = nodeSelector
+			set.Spec.Template.Spec.NodeSelector[quiesceNodeSelector] = quiesceNodeSelectorVal
 			err = client.Update(context.TODO(), &set)
 			if err != nil {
 				log.Trace(err)
@@ -475,7 +212,9 @@ func (t *Task) unQuiesceDaemonSets(client k8sclient.Client) error {
 }
 
 // Suspends all CronJobs
-func (t *Task) quiesceCronJobs(client k8sclient.Client) error {
+// Using unstructured because OCP 3.7 supports batch/v2alpha1 which is
+// not supported in newer versions such as OCP 3.11.
+func (t *Task) suspendCronJobs(client k8sclient.Client) error {
 	for _, ns := range t.sourceNamespaces() {
 		list := batchv1beta.CronJobList{}
 		options := k8sclient.InNamespace(ns)
@@ -485,13 +224,9 @@ func (t *Task) quiesceCronJobs(client k8sclient.Client) error {
 			return err
 		}
 		for _, r := range list.Items {
-			if r.Annotations == nil {
-				r.Annotations = make(map[string]string)
-			}
-			if r.Spec.Suspend == pointer.BoolPtr(true) {
+			if r.Spec.Suspend != nil && *r.Spec.Suspend {
 				continue
 			}
-			r.Annotations[SuspendAnnotation] = "true"
 			r.Spec.Suspend = pointer.BoolPtr(true)
 			err = client.Update(context.TODO(), &r)
 			if err != nil {
@@ -504,38 +239,8 @@ func (t *Task) quiesceCronJobs(client k8sclient.Client) error {
 	return nil
 }
 
-// Undo quiescence on all CronJobs
-func (t *Task) unQuiesceCronJobs(client k8sclient.Client) error {
-	for _, ns := range t.sourceNamespaces() {
-		list := batchv1beta.CronJobList{}
-		options := k8sclient.InNamespace(ns)
-		err := client.List(context.TODO(), options, &list)
-		if err != nil {
-			log.Trace(err)
-			return err
-		}
-		for _, r := range list.Items {
-			if r.Annotations == nil {
-				continue
-			}
-			if _, exist := r.Annotations[SuspendAnnotation]; !exist {
-				continue
-			}
-			delete(r.Annotations, SuspendAnnotation)
-			r.Spec.Suspend = pointer.BoolPtr(false)
-			err = client.Update(context.TODO(), &r)
-			if err != nil {
-				log.Trace(err)
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
 // Scales down all Jobs
-func (t *Task) quiesceJobs(client k8sclient.Client) error {
+func (t *Task) scaleDownJobs(client k8sclient.Client) error {
 	zero := int32(0)
 	for _, ns := range t.sourceNamespaces() {
 		list := batchv1.JobList{}
@@ -549,54 +254,10 @@ func (t *Task) quiesceJobs(client k8sclient.Client) error {
 			return err
 		}
 		for _, job := range list.Items {
-			if job.Annotations == nil {
-				job.Annotations = make(map[string]string)
-			}
 			if job.Spec.Parallelism == &zero {
 				continue
 			}
-			job.Annotations[ReplicasAnnotation] = strconv.FormatInt(int64(*job.Spec.Parallelism), 10)
 			job.Spec.Parallelism = &zero
-			err = client.Update(context.TODO(), &job)
-			if err != nil {
-				log.Trace(err)
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
-// Scales all Jobs back up
-func (t *Task) unQuiesceJobs(client k8sclient.Client) error {
-	for _, ns := range t.sourceNamespaces() {
-		list := batchv1.JobList{}
-		options := k8sclient.InNamespace(ns)
-		err := client.List(
-			context.TODO(),
-			options,
-			&list)
-		if err != nil {
-			log.Trace(err)
-			return err
-		}
-		for _, job := range list.Items {
-			if job.Annotations == nil {
-				continue
-			}
-			replicas, exist := job.Annotations[ReplicasAnnotation]
-			if !exist {
-				continue
-			}
-			number, err := strconv.Atoi(replicas)
-			if err != nil {
-				log.Trace(err)
-				return err
-			}
-			delete(job.Annotations, ReplicasAnnotation)
-			parallelReplicas := int32(number)
-			job.Spec.Parallelism = &parallelReplicas
 			err = client.Update(context.TODO(), &job)
 			if err != nil {
 				log.Trace(err)
@@ -640,9 +301,6 @@ func (t *Task) ensureQuiescedPodsTerminated() (bool, error) {
 			return false, err
 		}
 		for _, pod := range list.Items {
-			if pod.Annotations == nil {
-				pod.Annotations = make(map[string]string)
-			}
 			if _, found := skippedPhases[pod.Status.Phase]; found {
 				continue
 			}
